Define temporary path explicitly in jvm arg

+++ core/org.eclipse.emf.texo/src/org/eclipse/emf/texo/store/TexoEMFResourceURIConverter.java
@@ -60,6 +60,9 @@ public class TexoEMFResourceURIConverter extends ExtensibleURIConverterImpl impl
    */
   protected String getTemporaryDirectoryPath() {
     try {
+      if (System.getProperty("texo.temp.directory") != null) { //$NON-NLS-1$
+        return System.getProperty("texo.temp.directory");//$NON-NLS-1$
+      }
       final File f = File.createTempFile("test" + System.currentTimeMillis(), null); //$NON-NLS-1$
       String tempDirectory = f.getParentFile().getAbsolutePath();
       f.delete();
